extrach carousel into a separate component. add carouselSlide component for an easier, more ergonomic and intuitive implementation.
